put cache middleware to a separate function
