+++ package.json
@@ -46,6 +46,7 @@
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-prettier": "^4.2.1",
+    "morgan": "^1.10.0",
     "prettier": "^2.8.0",
     "supertest": "6.1.3",
     "tape": "5.2.2"
