+++ server.js
@@ -2,6 +2,7 @@ require('dotenv').config();
 
 const express = require('express');
 const cors = require('cors');
+const morgan = require('morgan');
 
 const apiRouter = require('./api-router.js');
 
@@ -16,16 +17,16 @@ if (permissiveCors) {
   };
 } else {
   corsOptions = {
-    origin: [
-      'https://rsk.co',
-      /\.rsk\.co$/,
-    ],
+    origin: ['https://rsk.co', /\.rsk\.co$/],
     optionsSuccessStatus: 200,
   };
 }
 
 server.use(cors(corsOptions));
 
+if (process.env.NODE_ENV != 'production') {
+  server.use(morgan('dev'));
+}
 server.get('/api/status', (req, res) => {
   res.send({
     ok: Date.now(),
