+++ util/activity-report/index.js
@@ -1,9 +1,7 @@
 const { queryAllActivity } = require('./all-activity.js');
 const { queryDeveloperActivity } = require('./developer-activity.js');
-const { dbQueryAvgTxCost } = require('./avg-tx-cost.js');
 
 module.exports = {
   queryAllActivity,
   queryDeveloperActivity,
-  dbQueryAvgTxCost,
 };
